chore/#81: ResponseBookmarkListDto 변수 타입 변경

+++ app/src/main/java/com/wearerommies/roomie/data/dto/response/ResponseBookmarkListDto.kt
@@ -12,7 +12,7 @@ data class ResponseBookmarkListDto(
     @Serializable
     data class PinnedHouse(
         @SerialName("contractTerm")
-        val contractTerm: Long,
+        val contractTerm: Int,
         @SerialName("deposit")
         val deposit: String,
         @SerialName("genderPolicy")
